requirement.txt
pytesseract
Pillow
python-docx
pdf2image
fitz  # (PyMuPDF for PDF handling)